changed /meme commands
